:truck: Renamed filename of NavigationListItem.tsx
