+++ packages/next-templates/src/components/ExampleFooter/ExampleFooter.tsx
@@ -1,7 +1,7 @@
 import Logo from '../../app/styling/assets/voorbeeld-footer.svg';
 import { NavigationList } from '../NavigationList/Navigationlist';
-import { NavigationListItem } from '../NavigationList/NavigationListIItem';
-import { Heading5, PageFooter, Paragraph } from '@utrecht/component-library-react';
+import { NavigationListItem } from '../NavigationList/NavigationListItem';
+import { Heading5, PageFooter, Paragraph, UnorderedList, UnorderedListItem } from '@utrecht/component-library-react';
 import React, { HTMLAttributes } from 'react';
 import './footer.css';
